refactor: flag creation button
