+++ frontend/src/component/project/Project/PaginatedProjectFeatureToggles/ProjectFeatureTogglesHeader/ProjectFeatureTogglesHeader.tsx
@@ -12,23 +12,14 @@ import { PageHeader } from 'component/common/PageHeader/PageHeader';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
 import { Search } from 'component/common/Search/Search';
 import { useUiFlag } from 'hooks/useUiFlag';
-import Add from '@mui/icons-material/Add';
-import { styled } from '@mui/material';
-import ResponsiveButton from 'component/common/ResponsiveButton/ResponsiveButton';
-import { useSearchParams } from 'react-router-dom';
 import { useRequiredPathParam } from 'hooks/useRequiredPathParam';
-import { CREATE_FEATURE } from 'component/providers/AccessProvider/permissions';
 import { ExportDialog } from 'component/feature/FeatureToggleList/ExportDialog';
 import type { FeatureSchema } from 'openapi';
 import { usePlausibleTracker } from 'hooks/usePlausibleTracker';
 import ReviewsOutlined from '@mui/icons-material/ReviewsOutlined';
 import { useFeedback } from 'component/feedbackNew/useFeedback';
-import useUiConfig from 'hooks/api/getters/useUiConfig/useUiConfig';
-import { CreateFeatureDialog } from './CreateFeatureDialog.tsx';
 import IosShare from '@mui/icons-material/IosShare';
-import type { OverridableStringUnion } from '@mui/types';
-import type { ButtonPropsVariantOverrides } from '@mui/material/Button/Button';
-import { NAVIGATE_TO_CREATE_FEATURE } from 'utils/testIds';
+import { FlagCreationButton } from './FlagCreationButton/FlagCreationButton.tsx';
 
 interface IProjectFeatureTogglesHeaderProps {
     isLoading?: boolean;
@@ -39,60 +30,6 @@ interface IProjectFeatureTogglesHeaderProps {
     environmentsToExport?: string[];
     actions?: ReactNode;
 }
-
-interface IFlagCreationButtonProps {
-    text?: string;
-    variant?: OverridableStringUnion<
-        'text' | 'outlined' | 'contained',
-        ButtonPropsVariantOverrides
-    >;
-    skipNavigationOnComplete?: boolean;
-    isLoading?: boolean;
-    onSuccess?: () => void;
-}
-
-const StyledResponsiveButton = styled(ResponsiveButton)(() => ({
-    whiteSpace: 'nowrap',
-}));
-
-export const FlagCreationButton = ({
-    variant,
-    text = 'New feature flag',
-    skipNavigationOnComplete,
-    isLoading,
-    onSuccess,
-}: IFlagCreationButtonProps) => {
-    const { loading } = useUiConfig();
-    const [searchParams] = useSearchParams();
-    const projectId = useRequiredPathParam('projectId');
-    const showCreateDialog = Boolean(searchParams.get('create'));
-    const [openCreateDialog, setOpenCreateDialog] = useState(showCreateDialog);
-
-    return (
-        <>
-            <StyledResponsiveButton
-                onClick={() => setOpenCreateDialog(true)}
-                maxWidth='960px'
-                Icon={Add}
-                projectId={projectId}
-                disabled={loading || isLoading}
-                variant={variant}
-                permission={CREATE_FEATURE}
-                data-testid={
-                    loading || isLoading ? '' : NAVIGATE_TO_CREATE_FEATURE
-                }
-            >
-                {text}
-            </StyledResponsiveButton>
-            <CreateFeatureDialog
-                open={openCreateDialog}
-                onClose={() => setOpenCreateDialog(false)}
-                skipNavigationOnComplete={skipNavigationOnComplete}
-                onSuccess={onSuccess}
-            />
-        </>
-    );
-};
 
 export const ProjectFeatureTogglesHeader: FC<
     IProjectFeatureTogglesHeaderProps
@@ -196,6 +133,8 @@ export const ProjectFeatureTogglesHeader: FC<
                                 />
                             }
                         />
+
+                        {/* FIXME: remove */}
                         <ConditionallyRender
                             condition={
                                 projectOverviewRefactorFeedback &&
